Renamed "ScrivitoExtentions" to "EditorTabs"
