+++ src/Components/EditorTabs/IconEditorTab/AllIcons.js
