+++ src/Components/EditorTabs/IconEditorTab/SingleIcon.js
